Refaktoryzacja do wzorca bridge, podpunkty 3.1.1. oraz 3.2.1
