+++ .idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="powp-rhymers" />
+        <module name="Wzorzec" />
       </profile>
     </annotationProcessing>
   </component>
